upd 29-6-19

+++ Math/Modular Multiplicative Inverse.cpp
@@ -1,19 +1,19 @@
-int extended_euclidean(int a, int b, int & x, int & y) {
+ll extended_euclidean(ll a, ll b, ll & x, ll & y) {
     if (a == 0) {
         x = 0;
         y = 1;
         return b;
     }
-    int x1, y1;
-    int d = extended_euclidean(b % a, a, x1, y1);
+    ll x1, y1;
+    ll d = extended_euclidean(b % a, a, x1, y1);
     x = y1 - (b / a) * x1;
     y = x1;
     return d;
 }
-int mod_inverse(int a,int m)
+ll mod_inverse(ll a,ll m)
 {
-    int x, y;
-    int g = extended_euclidean(a, m, x, y);
+    ll x, y;
+    ll g = extended_euclidean(a, m, x, y);
     if (g != 1)
     {
         cout << "No solution!";
@@ -25,9 +25,9 @@ int mod_inverse(int a,int m)
         return x;
     }
 }
-void for_all_mod_inverse(int m)
+void for_all_mod_inverse(ll m)
 {
     inv[1] = 1;
-    for(int i = 2; i < m; ++i)
+    for(ll i = 2; i < m; ++i)
         inv[i] = (m - (m/i) * inv[m%i] % m) % m;
 }
